# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisRef": {
        "additionalProperties": false,
        "description": "The readiness analysis an execution was created from (API and CLI require one).",
        "properties": {
          "aiAnalysisAvailable": {
            "title": "Aianalysisavailable",
            "type": "boolean"
          },
          "analyzedAt": {
            "format": "date-time",
            "title": "Analyzedat",
            "type": "string"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "riskLevel": {
            "title": "Risklevel",
            "type": "string"
          },
          "riskScore": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Riskscore",
            "type": "integer"
          }
        },
        "required": [
          "jobId",
          "riskScore",
          "riskLevel",
          "aiAnalysisAvailable",
          "analyzedAt"
        ],
        "title": "AnalysisRef",
        "type": "object"
      },
...
      "Approval": {
        "additionalProperties": false,
        "properties": {
          "approver": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Approver",
            "type": "string"
          },
          "callHash": {
            "title": "Callhash",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "verifyHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifyhash"
          }
        },
        "required": [
          "approver",
          "callHash",
          "createdAt"
        ],
        "title": "Approval",
        "type": "object"
      },
      "ApproveRequest": {
        "additionalProperties": false,
        "properties": {
          "approver": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Approver",
            "type": "string"
          },
          "callHash": {
            "description": "The hash the approver was shown",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Callhash",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "kind": {
            "default": "main",
            "enum": [
              "main",
              "rollback"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "approver",
          "callHash"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
...
      "CallKind": {
        "enum": [
          "main",
          "verify",
          "rollback"
        ],
        "title": "CallKind",
        "type": "string"
      },
      "CallRequest": {
        "additionalProperties": false,
        "description": "Replace a call, or accept it as it is (`call` omitted on the main call).",
        "properties": {
          "call": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannedToolCall"
              },
              {
                "type": "null"
              }
            ]
          },
          "editor": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Editor",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/CallKind",
            "default": "main"
          }
        },
        "required": [
          "editor"
        ],
        "title": "CallRequest",
        "type": "object"
      },
      "CallSource": {
        "enum": [
          "annotated",
          "ai_proposed",
          "edited"
        ],
        "title": "CallSource",
        "type": "string"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The shared error shape: `{error, code, details?}`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "Execution": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "auditHead": {
            "default": "",
            "title": "Audithead",
            "type": "string"
          },
          "auditSeq": {
            "default": 0,
            "title": "Auditseq",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "pauseReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pausereason"
          },
          "runbookLabel": {
            "title": "Runbooklabel",
            "type": "string"
          },
          "runbookSha256": {
            "title": "Runbooksha256",
            "type": "string"
          },
          "startedBy": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Startedby",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState",
            "default": "CREATED"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepRun"
            },
            "title": "Steps",
            "type": "array"
          },
          "toolCallsUsed": {
            "default": 0,
            "title": "Toolcallsused",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mode",
          "runbookLabel",
          "runbookSha256",
          "startedBy",
          "createdAt",
          "updatedAt",
          "steps"
        ],
        "title": "Execution",
        "type": "object"
      },
      "ExecutionMode": {
        "enum": [
          "dry_run",
          "live"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
...
      "ExecutionState": {
        "enum": [
          "CREATED",
          "RUNNING",
          "PAUSED",
          "COMPLETED",
          "FAILED",
          "ABORTED"
        ],
        "title": "ExecutionState",
        "type": "string"
      },
...
      "JsonValue": {},
...
      "PlannedToolCall": {
        "additionalProperties": false,
        "description": "A tool call written in a runbook annotation: `<server>/<tool> {JSON arguments}`.\n\nUntrusted input like any other runbook text: it is only a request, checked\nagainst the execution policy and approved by humans before anything runs.",
        "properties": {
          "arguments": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Arguments",
            "type": "object"
          },
          "server": {
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Server",
            "type": "string"
          },
          "tool": {
            "pattern": "^[A-Za-z0-9_.-]{1,128}$",
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "server",
          "tool"
        ],
        "title": "PlannedToolCall",
        "type": "object"
      },
      "ProposalNote": {
        "additionalProperties": false,
        "description": "Why the model proposed a step's call; untrusted text, shown to reviewers as-is.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "rationale": {
            "default": "",
            "maxLength": 1000,
            "title": "Rationale",
            "type": "string"
          }
        },
        "title": "ProposalNote",
        "type": "object"
      },
...
      "RiskClass": {
        "enum": [
          "read",
          "write",
          "destructive"
        ],
        "title": "RiskClass",
        "type": "string"
      },
...
      "StepDecisionRequest": {
        "additionalProperties": false,
        "description": "`reject`, `skip` and `manual` need a reason; `verify` needs `succeeded`.",
        "properties": {
          "actor": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "succeeded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Succeeded"
          }
        },
        "required": [
          "actor"
        ],
        "title": "StepDecisionRequest",
        "type": "object"
      },
...
      "StepRun": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "attempt": {
            "default": 1,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "call": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolCall"
              },
              {
                "type": "null"
              }
            ]
          },
          "dependsOn": {
            "items": {
              "type": "integer"
            },
            "title": "Dependson",
            "type": "array"
          },
          "estimatedMinutes": {
            "title": "Estimatedminutes",
            "type": "number"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "phase": {
            "minimum": 1.0,
            "title": "Phase",
            "type": "integer"
          },
          "policyErrors": {
            "items": {
              "type": "string"
            },
            "title": "Policyerrors",
            "type": "array"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalNote"
              },
              {
                "type": "null"
              }
            ]
          },
          "rollback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolCall"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/StepState",
            "default": "PLANNED"
          },
          "stepNumber": {
            "minimum": 1.0,
            "title": "Stepnumber",
            "type": "integer"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "targetSystem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetsystem"
          },
          "verify": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolCall"
              },
              {
                "type": "null"
              }
            ]
          },
          "waitingSince": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waitingsince"
          }
        },
        "required": [
          "stepNumber",
          "phase",
          "action",
          "owner",
          "estimatedMinutes"
        ],
        "title": "StepRun",
        "type": "object"
      },
      "StepState": {
        "enum": [
          "PLANNED",
          "PROPOSED",
          "AWAITING_APPROVAL",
          "APPROVED",
          "REJECTED",
          "RUNNING",
          "VERIFYING",
          "SUCCEEDED",
          "FAILED",
          "UNKNOWN",
          "ROLLED_BACK",
          "AWAITING_MANUAL",
          "MANUAL_DONE",
          "SKIPPED"
        ],
        "title": "StepState",
        "type": "string"
      },
...
      "ToolCall": {
        "additionalProperties": false,
        "properties": {
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/Approval"
            },
            "title": "Approvals",
            "type": "array"
          },
          "arguments": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Arguments",
            "type": "object"
          },
          "callHash": {
            "title": "Callhash",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "kind": {
            "$ref": "#/components/schemas/CallKind"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "riskClass": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "server": {
            "title": "Server",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/CallSource"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "server",
          "tool",
          "source",
          "callHash"
        ],
        "title": "ToolCall",
        "type": "object"
      },
      "ToolResult": {
        "additionalProperties": false,
        "description": "Outcome of one call. `ok=False` is a tool-level failure reported by the server.",
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "simulated": {
            "default": false,
            "title": "Simulated",
            "type": "boolean"
          },
          "structured": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ToolResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Parse DR runbooks, check dependency health and produce readiness reports.",
    "title": "DR Readiness Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/executions/{execution_id}/steps/{step_number}/approve": {
      "post": {
        "description": "Approve the call with the hash you were shown (`409 STALE_CALL` if it changed).",
        "operationId": "approve_api_v1_executions__execution_id__steps__step_number__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_number",
            "required": true,
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "title": "Step Number",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Approve",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/steps/{step_number}/call": {
      "put": {
        "description": "Accept (omit `call`) or replace a call. Clears the step's approvals.",
        "operationId": "set_call_api_v1_executions__execution_id__steps__step_number__call_put",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_number",
            "required": true,
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "title": "Step Number",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Set Call",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/steps/{step_number}/{action}": {
      "post": {
        "operationId": "decide_api_v1_executions__execution_id__steps__step_number___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_number",
            "required": true,
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "title": "Step Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "reject",
                "skip",
                "manual",
                "mark-done",
                "verify",
                "retry",
                "rollback"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Decide",
        "tags": [
          "execution"
        ]
      }
    },
...
  }
}